+++ arrow-libs/core/arrow-continuations/build.gradle
@@ -12,8 +12,21 @@ apply plugin: 'kotlinx-atomicfu'
 
 kotlin {
     js(BOTH) {
-        browser()
-        nodejs()
+        browser {
+            testTask {
+                useKarma {
+                    useChromeHeadless()
+                    useConfigDirectory("${project.rootDir}/karma")
+                }
+            }
+        }
+        nodejs {
+            testTask {
+                useMocha {
+                    timeout = "10000000"
+                }
+            }
+        }
     }
     sourceSets {
         commonMain {
@@ -21,14 +34,29 @@ kotlin {
                 compileOnly "org.jetbrains.kotlin:kotlin-stdlib-common:$KOTLIN_VERSION"
             }
         }
+        commonTest {
+            dependencies {
+                implementation project(":arrow-core-test")
+            }
+        }
         jvmMain {
             dependencies {
                 compileOnly "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$KOTLIN_VERSION"
             }
         }
+        jvmTest {
+            dependencies {
+                runtimeOnly "io.kotest:kotest-runner-junit5:$KOTEST_VERSION"
+            }
+        }
         jsMain {
             dependencies {
                 compileOnly "org.jetbrains.kotlin:kotlin-stdlib-js:$KOTLIN_VERSION"
+            }
+        }
+        jsTest {
+            dependencies {
+                implementation "io.kotest:kotest-framework-engine:$KOTEST_VERSION"
             }
         }
     }
