+++ arrow-libs/core/arrow-continuations/src/commonTest/kotlin/generic/MaybeEffect.kt
@@ -1,0 +1,11 @@
+package generic
+
+import arrow.continuations.Effect
+
+fun interface MaybeEffect<A> : Effect<Maybe<A>> {
+  suspend fun <B> Maybe<B>.bind(): B =
+    when (this) {
+      is Just -> a
+      is None -> control().shift(None)
+    }
+}
