+++ arrow-libs/core/arrow-core-retrofit/src/test/kotlin/arrow/retrofit/adapter/mock/ErrorMock.kt
