+++ arrow-libs/core/arrow-core-test/build.gradle
@@ -10,22 +10,8 @@ apply plugin: 'kotlinx-atomicfu'
 
 kotlin {
     js(BOTH) {
-        browser {
-            testTask {
-                useKarma {
-                    useChromeHeadless()
-                    println("${project.projectDir}/karma")
-                    useConfigDirectory("${project.projectDir}/karma")
-                }
-            }
-        }
-        nodejs {
-            testTask {
-                useMocha {
-                    timeout = "10000"
-                }
-            }
-        }
+        browser()
+        nodejs()
     }
     sourceSets {
         commonMain {
