+++ arrow-libs/core/arrow-core-test/src/commonMain/kotlin/arrow/core/test/Platform.kt
@@ -1,4 +1,4 @@
-package arrow.core
+package arrow.core.test
 
 expect fun isJvm(): Boolean
 expect fun isJs(): Boolean
