+++ arrow-libs/core/arrow-core-test/src/jsMain/kotlin/arrow.core/test/Platform-js.kt
@@ -1,4 +1,4 @@
-package arrow.core
+package arrow.core.test
 
 actual fun isJvm(): Boolean = false
 actual fun isJs(): Boolean = true
