+++ arrow-libs/core/arrow-core-test/src/jvmMain/kotlin/arrow/core/test/Platform-jvm.kt
@@ -1,4 +1,4 @@
-package arrow.core
+package arrow.core.test
 
 actual fun isJvm(): Boolean = true
 actual fun isJs(): Boolean = false
