+++ arrow-libs/core/arrow-core/build.gradle
@@ -13,8 +13,21 @@ apply from: "$ANIMALSNIFFER_MPP"
 
 kotlin {
     js(BOTH) {
-        browser()
-        nodejs()
+        browser {
+            testTask {
+                useKarma {
+                    useChromeHeadless()
+                    useConfigDirectory("${project.rootDir}/karma")
+                }
+            }
+        }
+        nodejs {
+            testTask {
+                useMocha {
+                    timeout = "10000000"
+                }
+            }
+        }
     }
     sourceSets {
         commonMain {
@@ -24,14 +37,29 @@ kotlin {
                 api project(":arrow-annotations")
             }
         }
+        commonTest {
+            dependencies {
+                implementation project(":arrow-core-test")
+            }
+        }
         jvmMain {
             dependencies {
                 compileOnly "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$KOTLIN_VERSION"
             }
         }
+        jvmTest {
+            dependencies {
+                runtimeOnly "io.kotest:kotest-runner-junit5:$KOTEST_VERSION"
+            }
+        }
         jsMain {
             dependencies {
                 compileOnly "org.jetbrains.kotlin:kotlin-stdlib-js:$KOTLIN_VERSION"
+            }
+        }
+        jsTest {
+            dependencies {
+                implementation "io.kotest:kotest-framework-engine:$KOTEST_VERSION"
             }
         }
     }
