+++ arrow-libs/core/arrow-core/src/commonTest/kotlin/arrow/core/ComparisonKtTest.kt
@@ -1,5 +1,6 @@
 package arrow.core
 
+import arrow.core.test.generators.array
 import io.kotest.assertions.assertSoftly
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.property.Arb
