+++ arrow-libs/core/arrow-core/src/commonTest/kotlin/arrow/core/EitherTest.kt
