+++ arrow-libs/core/arrow-core/src/jvmTest/java/arrow/typeclasses/SemiringUsageTest.java
